Prevent stale computations from being run
Add a 'triggersExit' function to all admin routes
that switches the currently displayed content to
the loading screen.

The Tracker is then flushed to update the view,
evict the current content and force it to stop
its computations.

This allows to change the session variables in
the 'tiggersEnter' and 'action' functions without
running computations not prepared to work with the
new values.

Fixes #19